requirements.txt
flask
pymongo
dnspython
python-dotenv
boto3
pytest
